Factor out local_pv_roof
