+++ src/climatevision/generator/electricity2030/electricity2030_commune.py
@@ -23,114 +23,8 @@ from .electricity2030_core import (
     Energy,
     calc_production_renewable_geothermal,
     calc_stop_production_by_fossil_fuels,
+    calc_production_local_pv_roof,
 )
-
-
-def calc_production_local_pv_roof(
-    inputs: Inputs,
-    *,
-    e18: E18,
-    b18: B18,
-    r18: R18,
-):
-    entries = inputs.entries
-    ass = inputs.ass
-    Kalkulationszeitraum = entries.m_duration_target
-
-    # TODO: Change the below
-    p_local_pv_roof = EColVars2030()
-
-    p_local_pv_roof.full_load_hour = entries.e_pv_full_load_hours_sta
-    p_local_pv_roof.power_installed = entries.e_PV_power_inst_roof
-    p_local_pv_roof.invest_per_x = (
-        ass("Ass_E_P_local_pv_roof_ratio_invest_to_power_2030") * 1000
-    )
-    p_local_pv_roof.pct_of_wage = ass("Ass_E_P_pv_invest_pct_of_wage")
-    p_local_pv_roof.ratio_wage_to_emplo = ass(
-        "Ass_E_P_constr_elec_ratio_wage_to_emplo_2017"
-    )
-    p_local_pv_roof.power_to_be_installed_pct = entries.e_PV_power_to_be_inst_roof
-    p_local_pv_roof.area_ha_available = (
-        (4 / 3)
-        * (
-            (
-                entries.r_area_m2_1flat
-                / 100
-                * ass("Ass_E_P_local_pv_roof_area_building1")
-                + entries.r_area_m2_2flat
-                / 100
-                * ass("Ass_E_P_local_pv_roof_area_building2")
-                + entries.r_area_m2_3flat
-                / 100
-                * ass("Ass_E_P_local_pv_roof_area_buildingD")
-                + entries.r_area_m2_dorm
-                / 100
-                * ass("Ass_E_P_local_pv_roof_area_buildingD")
-            )
-        )
-        / 10000
-    )
-    p_local_pv_roof.area_ha_available_pct_of_action = ass(
-        "Ass_E_P_local_pv_roof_potential"
-    )
-    p_local_pv_roof.ratio_power_to_area_ha = ass(
-        "Ass_E_P_local_pv_roof_ratio_power_to_area_ha"
-    )
-    p_local_pv_roof.power_installable = (
-        p_local_pv_roof.area_ha_available
-        * p_local_pv_roof.area_ha_available_pct_of_action
-        * p_local_pv_roof.ratio_power_to_area_ha
-    )
-    p_local_pv_roof.power_to_be_installed = max(
-        0,
-        p_local_pv_roof.power_installable * p_local_pv_roof.power_to_be_installed_pct
-        - p_local_pv_roof.power_installed,
-    )
-    p_local_pv_roof.energy_installable = (
-        p_local_pv_roof.power_installable
-        * p_local_pv_roof.full_load_hour
-        * (1 - ass("Ass_E_P_renew_loss_brutto_to_netto"))
-    )
-    p_local_pv_roof.cost_mro_per_MWh = (
-        ass("Ass_E_P_local_pv_roof_ratio_invest_to_power_2030")
-        * ass("Ass_E_P_local_pv_roof_mro_per_year")
-        / p_local_pv_roof.full_load_hour
-        * 1000
-    )
-    p_local_pv_roof.energy = (
-        (p_local_pv_roof.power_to_be_installed + p_local_pv_roof.power_installed)
-        * p_local_pv_roof.full_load_hour
-        * (1 - ass("Ass_E_P_renew_loss_brutto_to_netto"))
-    )
-    p_local_pv_roof.invest = (
-        p_local_pv_roof.power_to_be_installed * p_local_pv_roof.invest_per_x
-    )
-    p_local_pv_roof.cost_mro = (
-        p_local_pv_roof.energy * p_local_pv_roof.cost_mro_per_MWh / MILLION
-    )
-    p_local_pv_roof.change_energy_MWh = (
-        p_local_pv_roof.energy - e18.p_local_pv_roof.energy
-    )
-    p_local_pv_roof.invest_pa = p_local_pv_roof.invest / Kalkulationszeitraum
-    p_local_pv_roof.invest_com = div(
-        p_local_pv_roof.invest
-        * (r18.p_buildings_area_m2_com.area_m2 + b18.p_nonresi_com.area_m2),
-        b18.p_nonresi.area_m2 + r18.p_buildings_total.area_m2,
-    )
-    p_local_pv_roof.change_cost_mro = (
-        p_local_pv_roof.cost_mro - e18.p_local_pv_roof.cost_mro
-    )
-    p_local_pv_roof.change_energy_pct = div(
-        p_local_pv_roof.change_energy_MWh, e18.p_local_pv_roof.energy
-    )
-    p_local_pv_roof.cost_wage = p_local_pv_roof.invest_pa * p_local_pv_roof.pct_of_wage
-    p_local_pv_roof.invest_pa_com = p_local_pv_roof.invest_com / Kalkulationszeitraum
-    p_local_pv_roof.change_CO2e_t = 0
-    p_local_pv_roof.cost_climate_saved = 0
-    p_local_pv_roof.CO2e_total = 0
-    p_local_pv_roof.change_CO2e_pct = 0
-
-    return p_local_pv_roof
 
 
 def calc_production_local_pv_facade(
