+++ src/climatevision/generator/electricity2030/electricity2030_germany.py
@@ -23,6 +23,7 @@ from .electricity2030_core import (
     Energy,
     calc_production_renewable_geothermal,
     calc_stop_production_by_fossil_fuels,
+    calc_production_local_pv_roof,
 )
 
 
@@ -83,7 +84,7 @@ def calc(
     p_renew_reverse = EColVars2030()
     p_local = EColVars2030()
     p_local_pv = EColVars2030()
-    p_local_pv_roof = EColVars2030()
+    p_local_pv_roof = calc_production_local_pv_roof(inputs, e18=e18, b18=b18, r18=r18)
     p_local_pv_facade = EColVars2030()
     p_local_pv_park = EColVars2030()
     p_local_pv_agri = EColVars2030()
@@ -183,7 +184,6 @@ def calc(
     p_renew_pv.CO2e_combustion_based_per_MWh = fact(
         "Fact_E_P_climate_neutral_ratio_CO2e_cb_to_fec"
     )
-    p_local_pv_roof.full_load_hour = entries.e_pv_full_load_hours_sta
     p_local_pv_facade.full_load_hour = ass("Ass_E_P_local_pv_facade_full_load_hours")
     p_renew_wind.CO2e_combustion_based_per_MWh = fact(
         "Fact_E_P_climate_neutral_ratio_CO2e_cb_to_fec"
@@ -263,44 +263,9 @@ def calc(
         * entries.m_population_com_2018
         / entries.m_population_nat
     )
-    p_local_pv_roof.power_installed = entries.e_PV_power_inst_roof
     p_local_pv_facade.power_installed = entries.e_PV_power_inst_facade
     p_local_pv_park.power_installed = entries.e_PV_power_inst_park
     p_local_pv_agri.power_installed = entries.e_PV_power_inst_agripv
-    p_local_pv_roof.invest_per_x = (
-        ass("Ass_E_P_local_pv_roof_ratio_invest_to_power_2030") * 1000
-    )
-    p_local_pv_roof.pct_of_wage = ass("Ass_E_P_pv_invest_pct_of_wage")
-    p_local_pv_roof.ratio_wage_to_emplo = ass(
-        "Ass_E_P_constr_elec_ratio_wage_to_emplo_2017"
-    )
-    p_local_pv_roof.power_to_be_installed_pct = entries.e_PV_power_to_be_inst_roof
-    p_local_pv_roof.area_ha_available = (
-        (4 / 3)
-        * (
-            (
-                entries.r_area_m2_1flat
-                / 100
-                * ass("Ass_E_P_local_pv_roof_area_building1")
-                + entries.r_area_m2_2flat
-                / 100
-                * ass("Ass_E_P_local_pv_roof_area_building2")
-                + entries.r_area_m2_3flat
-                / 100
-                * ass("Ass_E_P_local_pv_roof_area_buildingD")
-                + entries.r_area_m2_dorm
-                / 100
-                * ass("Ass_E_P_local_pv_roof_area_buildingD")
-            )
-        )
-        / 10000
-    )
-    p_local_pv_roof.area_ha_available_pct_of_action = ass(
-        "Ass_E_P_local_pv_roof_potential"
-    )
-    p_local_pv_roof.ratio_power_to_area_ha = ass(
-        "Ass_E_P_local_pv_roof_ratio_power_to_area_ha"
-    )
     p_local_pv_facade.cost_mro_per_MWh = (
         ass("Ass_E_S_local_pv_facade_ratio_invest_to_power")
         * ass("Ass_E_P_local_pv_roof_mro_per_year")
@@ -485,12 +450,6 @@ def calc(
         / p_local_pv_roof.full_load_hour
         * 1000
     )
-    p_local_pv_roof.cost_mro_per_MWh = (
-        ass("Ass_E_P_local_pv_roof_ratio_invest_to_power_2030")
-        * ass("Ass_E_P_local_pv_roof_mro_per_year")
-        / p_local_pv_roof.full_load_hour
-        * 1000
-    )
     p_local_pv_park.full_load_hour = p_local_pv_roof.full_load_hour
     p_local_pv_agri.full_load_hour = p_local_pv_roof.full_load_hour
     p_renew_pv_facade.energy = 0
@@ -534,11 +493,6 @@ def calc(
         + p_local_pv_park.power_installed
         + p_local_pv_agri.power_installed
     )  #
-    p_local_pv_roof.power_installable = (
-        p_local_pv_roof.area_ha_available
-        * p_local_pv_roof.area_ha_available_pct_of_action
-        * p_local_pv_roof.ratio_power_to_area_ha
-    )
     p_local_pv_facade.power_installable = (
         p_local_pv_facade.ratio_power_to_area_ha
         * p_local_pv_facade.area_ha_available
@@ -676,16 +630,6 @@ def calc(
     p_renew_reverse.invest = (
         p_renew_reverse.power_to_be_installed * p_renew_reverse.invest_per_x
     )
-    p_local_pv_roof.power_to_be_installed = max(
-        0,
-        p_local_pv_roof.power_installable * p_local_pv_roof.power_to_be_installed_pct
-        - p_local_pv_roof.power_installed,
-    )
-    p_local_pv_roof.energy_installable = (
-        p_local_pv_roof.power_installable
-        * p_local_pv_roof.full_load_hour
-        * (1 - ass("Ass_E_P_renew_loss_brutto_to_netto"))
-    )
     p_local_pv_facade.power_to_be_installed = max(
         0,
         p_local_pv_facade.power_installable
@@ -819,14 +763,6 @@ def calc(
     )
 
     p_renew_reverse.invest_pa = p_renew_reverse.invest / Kalkulationszeitraum
-    p_local_pv_roof.energy = (
-        (p_local_pv_roof.power_to_be_installed + p_local_pv_roof.power_installed)
-        * p_local_pv_roof.full_load_hour
-        * (1 - ass("Ass_E_P_renew_loss_brutto_to_netto"))
-    )
-    p_local_pv_roof.invest = (
-        p_local_pv_roof.power_to_be_installed * p_local_pv_roof.invest_per_x
-    )
     p_local_pv_facade.energy = (
         (p_local_pv_facade.power_to_be_installed + p_local_pv_facade.power_installed)
         * p_local_pv_facade.full_load_hour
@@ -926,18 +862,6 @@ def calc(
     p_renew_hydro.energy = 0
     p_renew_reverse.cost_wage = p_renew_reverse.invest_pa * p_renew_reverse.pct_of_wage
 
-    p_local_pv_roof.cost_mro = (
-        p_local_pv_roof.energy * p_local_pv_roof.cost_mro_per_MWh / MILLION
-    )
-    p_local_pv_roof.change_energy_MWh = (
-        p_local_pv_roof.energy - e18.p_local_pv_roof.energy
-    )
-    p_local_pv_roof.invest_pa = p_local_pv_roof.invest / Kalkulationszeitraum
-    p_local_pv_roof.invest_com = (
-        p_local_pv_roof.invest
-        * (r18.p_buildings_area_m2_com.area_m2 + b18.p_nonresi_com.area_m2)
-        / (b18.p_nonresi.area_m2 + r18.p_buildings_total.area_m2)
-    )
     p_local_pv_facade.cost_mro = (
         p_local_pv_facade.energy * p_local_pv_facade.cost_mro_per_MWh / MILLION
     )
@@ -1051,14 +975,6 @@ def calc(
     p_renew_reverse.demand_emplo = div(
         p_renew_reverse.cost_wage, p_renew_reverse.ratio_wage_to_emplo
     )
-    p_local_pv_roof.change_cost_mro = (
-        p_local_pv_roof.cost_mro - e18.p_local_pv_roof.cost_mro
-    )
-    p_local_pv_roof.change_energy_pct = div(
-        p_local_pv_roof.change_energy_MWh, e18.p_local_pv_roof.energy
-    )
-    p_local_pv_roof.cost_wage = p_local_pv_roof.invest_pa * p_local_pv_roof.pct_of_wage
-    p_local_pv_roof.invest_pa_com = p_local_pv_roof.invest_com / Kalkulationszeitraum
     p_local_pv_facade.change_cost_mro = (
         p_local_pv_facade.cost_mro - e18.p_local_pv_facade.cost_mro
     )
@@ -1568,8 +1484,8 @@ def calc(
     p_local_pv_facade.cost_climate_saved = 0
     p_local_pv_agri.change_CO2e_t = 0
     p_local_pv_agri.cost_climate_saved = 0
-    p_local_pv_roof.change_CO2e_t = 0
-    p_local_pv_roof.cost_climate_saved = 0
+    # p_local_pv_roof.change_CO2e_t = 0
+    # p_local_pv_roof.cost_climate_saved = 0
 
     p_local_wind_onshore.CO2e_total = 0
     p_local_wind_onshore.cost_climate_saved = 0
@@ -1627,7 +1543,6 @@ def calc(
 
     p_local.change_CO2e_pct = 0
     p_local_pv.change_CO2e_pct = 0
-    p_local_pv_roof.change_CO2e_pct = 0
     p_local_pv_agri.change_CO2e_pct = 0
     p_local_pv_facade.change_CO2e_pct = 0
     p_local_pv_park.change_CO2e_pct = 0
@@ -1652,7 +1567,6 @@ def calc(
 
     p_local.cost_climate_saved = 0
     p_local_pv.cost_climate_saved = 0
-    p_local_pv_roof.cost_climate_saved = 0
     p_local_pv_agri.cost_climate_saved = 0
     p_local_pv_facade.cost_climate_saved = 0
     p_local_pv_park.cost_climate_saved = 0
